Updated controllers for deleting and posting comments.

+++ src/main/java/edu/cnm/deepdive/funrun/controller/HistoryController.java
@@ -1,18 +1,20 @@
 package edu.cnm.deepdive.funrun.controller;
 
-import edu.cnm.deepdive.funrun.model.entity.Comment;
 import edu.cnm.deepdive.funrun.model.entity.History;
 import edu.cnm.deepdive.funrun.model.entity.User;
+import edu.cnm.deepdive.funrun.model.entity.User.Role;
 import edu.cnm.deepdive.funrun.service.CommentRepository;
 import edu.cnm.deepdive.funrun.service.HistoryRepository;
 import edu.cnm.deepdive.funrun.service.UserRepository;
-import java.util.List;
+import edu.cnm.deepdive.funrun.service.UserService;
+import java.security.AccessControlException;
 import java.util.NoSuchElementException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.hateoas.server.ExposesResourceFor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -33,22 +35,24 @@ import org.springframework.web.bind.annotation.RestController;
 public class HistoryController {
 
   private final HistoryRepository historyRepository;
-  private final UserRepository userRepository;
   private final CommentRepository commentRepository;
+  private final UserService userService;
 
   /**
    * Allows histories to be posted, retrieved, updated, and deleted.
    *
    * @param historyRepository, commentRepository, userRepository
+   * @param userService
    * @return histories.
    */
   @Autowired
   public HistoryController(HistoryRepository historyRepository,
-      UserRepository userRepository, CommentRepository commentRepository) {
+      CommentRepository commentRepository,
+      UserService userService) {
     this.historyRepository = historyRepository;
-    this.userRepository = userRepository;
     this.commentRepository = commentRepository;
 
+    this.userService = userService;
   }
   /**
    * Allows histories to be retrieved.
@@ -57,8 +61,9 @@ public class HistoryController {
    * @return users ordered by name ascending.
    */
   @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
-  public Iterable<User> get() {
-    return userRepository.getAllByOrderByNameAsc();
+  public Iterable<History> get(Authentication auth) {
+    User user = userService.getCurrentUser(auth);
+    return historyRepository.getAllByUserOrderByStartDescEndDesc(user);
   }
 
   /**
@@ -69,7 +74,9 @@ public class HistoryController {
    */
   @PostMapping(
       consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
-  public ResponseEntity<History> post(@RequestBody History history) {
+  public ResponseEntity<History> post(@RequestBody History history, Authentication auth) {
+    User user = userService.getCurrentUser(auth);
+    history.setUser(user);
     historyRepository.save(history);
     return ResponseEntity.created(history.getHref()).body(history);
   }
@@ -94,14 +101,20 @@ public class HistoryController {
   @PutMapping(value = "/{id:\\d+}",
       consumes = MediaType.APPLICATION_JSON_VALUE,
       produces = MediaType.APPLICATION_JSON_VALUE)
-  public History put(@PathVariable long id, @RequestBody History history) {
-    History existingHistory = get(id);
-    if (history.getStart() != null) {
-      existingHistory.setStart(
-          history.getStart());
-    }
-    return historyRepository.save(existingHistory);
+  public History put(@PathVariable long id, @RequestBody History history, Authentication auth) {
+    User user = userService.getCurrentUser(auth);
+    return historyRepository.findById(id)
+        .map((existing) -> {
+          userService.requireAccess(user, existing.getUser().getId());
+          if (history.getStart() != null) {
+            existing.setStart(
+                history.getStart());
+          }
+          return historyRepository.save(existing);
+        })
+        .orElseThrow(NoSuchElementException::new);
   }
+
 
   /**
    * Allows histories to be deleted.
@@ -111,21 +124,17 @@ public class HistoryController {
    */
   @DeleteMapping(value = "/{id:\\d+}")
   @ResponseStatus(HttpStatus.NO_CONTENT)
-  public void delete(@PathVariable long id) {
+  public void delete(@PathVariable long id, Authentication auth) {
+    User user = userService.getCurrentUser(auth);
     historyRepository.findById(id)
         .map((history) -> {
-          List<Comment> comments = history.getComments();
-          comments.forEach(
-              (comment) -> comment.setHistory(null));
-          commentRepository.saveAll(comments);
-          return history;
-        })
-        .map((history) -> {
+          userService.requireAccess(user, history.getUser().getId());
           historyRepository.delete(history);
-          return history;
+          return null;
         })
-        .orElseThrow(NoSuchElementException::new);
+        .orElse(null);
 
   }
+
 }
